Create FoodSErviceImpl

+++ backend-spring-boot/Online Food Ordering/Online-Food-Ordering/src/main/java/com/heli/repository/FoodRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface FoodRepository extends JpaRepository<Food, Integer> {
 
 
-    List<Food> findByRestaurantId(Long restaurantaId);
+    List<Food> findByRestaurantId(int restaurantaId);
 
     @Query("SELECT f FROM Food f WHERE f.name LIKE %:keyword% OR f.foodCategory.name LIKE %:keyword%")
     List<Food> searchFood(@Param("keyword") String keyword);
